DS-354: refactor form examples to their own files
